Feito a restruturação do projeto
Foi realocado todas as pastas, para que o projeto fique organizado. Além de ter reogarnizado todas as pastas foi instalado Redux.
