+++ Aplicacao/package.json
@@ -16,8 +16,10 @@
     "react-google-charts": "^3.0.15",
     "react-icons": "^4.1.0",
     "react-leaflet": "^3.1.0",
+    "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
+    "redux": "^4.0.5",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
