+++ Aplicacao/src/Connections/api.js
