+++ Aplicacao/src/Pages/Indicator/index.js
