+++ Aplicacao/src/Pages/Indicator/styles.css
