--- Aplicacao/src/components/Menu/styles.css
@@ -1,6 +1,0 @@
-.link{
-    text-decoration: none;
-    margin-left: 5%;
-    color: rgb(129, 129, 129);
-    
-}
