+++ Aplicacao/src/rotas.js
@@ -2,16 +2,15 @@ import React from 'react';
 import {Switch, Route} from 'react-router-dom';
 
 
-import Home from '../components/Pages/Home';
-import Tabela from '../components/Pages/Table';
-import Indicador from '../components/Pages/Indicator';
-import Graph from '../components/Pages/Graph';
-import Mapa from '../components/Pages/Map';
-import Cadastro from '../components/Pages/Cadastro';
+import Home from './Pages/Home';
+import Tabela from './Pages/Table';
+import Indicador from './Pages/Indicator';
+import Graph from './Pages/Graph';
+import Mapa from './Pages/Map';
+import Cadastro from './Pages/Cadastro';
 
 export default function Router(){
     return(
-        
         <Switch>
             <Route path="/" exact component={Home}/>
             <Route path="/tabela" component={Tabela}/>
